Address review comments by @ishan16696

+++ internal/operator/statefulset/stsmatcher.go
@@ -268,7 +268,7 @@ func (s StatefulSetMatcher) matchEtcdContainerReadinessHandler() gomegatypes.Gom
 	}
 	scheme := utils.IfConditionOr(s.etcd.Spec.Backup.TLS == nil, corev1.URISchemeHTTP, corev1.URISchemeHTTPS)
 	path := utils.IfConditionOr(s.etcd.Spec.Replicas > 1, "/readyz", "/healthz")
-	port := utils.IfConditionOr(s.etcd.Spec.Replicas > 1, 9095, 8080)
+	port := utils.IfConditionOr(s.etcd.Spec.Replicas > 1, int32(9095), int32(8080))
 	return MatchFields(IgnoreExtras|IgnoreMissing, Fields{
 		"HTTPGet": PointTo(MatchFields(IgnoreExtras|IgnoreMissing, Fields{
 			"Path": Equal(path),
